create command decorator change exitMessage
