+++ src/main/java/com/andersenlab/hotel/application/command/BusinessExceptionHandlingCommand.java
@@ -10,9 +10,6 @@ import java.io.PrintStream;
 import java.util.List;
 
 public final class BusinessExceptionHandlingCommand implements Command {
-
-    private static final ApplicationCommand applicationCommand = ApplicationCommand.BUSINESS_EXCEPTION_HANDLER;
-
     private final Command original;
 
     public BusinessExceptionHandlingCommand(Command original) {
@@ -21,7 +18,7 @@ public final class BusinessExceptionHandlingCommand implements Command {
 
     @Override
     public ApplicationCommand getApplicationCommand() {
-        return applicationCommand;
+        return original.getApplicationCommand(); //instead of return applicationCommand;
     }
 
     @Override
