+++ src/main/java/com/andersenlab/hotel/application/command/GenericExceptionHandlingCommand.java
@@ -6,9 +6,6 @@ import java.io.PrintStream;
 import java.util.List;
 
 public final class GenericExceptionHandlingCommand implements Command {
-
-    private static final ApplicationCommand applicationCommand = ApplicationCommand.GENERIC_EXCEPTION_HANDLER;
-
     private final Command original;
 
     public GenericExceptionHandlingCommand(Command original) {
@@ -17,7 +14,7 @@ public final class GenericExceptionHandlingCommand implements Command {
 
     @Override
     public ApplicationCommand getApplicationCommand() {
-        return applicationCommand;
+        return original.getApplicationCommand(); //instead of return applicationCommand;
     }
 
     @Override
