+++ src/main/java/com/andersenlab/hotel/application/command/additional/ExitApplicationCommand.java
@@ -7,8 +7,6 @@ import java.io.PrintStream;
 import java.util.List;
 
 public final class ExitApplicationCommand implements Command {
-
-    private static final String EXIT_MESSAGE = "Goodbye, have a nice day!";
     private static final ApplicationCommand APPLICATION_COMMAND = ApplicationCommand.EXIT;
 
     @Override
@@ -18,6 +16,6 @@ public final class ExitApplicationCommand implements Command {
 
     @Override
     public void execute(PrintStream output, List<String> arguments) {
-        output.println(EXIT_MESSAGE);
+        output.println("Goodbye, have a nice day!");
     }
 }
